requirements.txt
Pillow==9.0.1
Unidecode==1.3.6
yt-dlp==2023.3.4
